hlt-7: some refactoring
