+++ scripts/pretest.sh
@@ -1,13 +1,9 @@
 #!/usr/bin/env bash
 
-echo 'Linting sources'
-find ./src -name '*.ts' -print0 | xargs -0 node_modules/tslint/bin/tslint -c ./tslint.json || exit 1
+sh ./scripts/compile.sh || exit 1
 
 echo 'Compiling karma.conf.ts'
 node_modules/typescript/bin/tsc karma.conf.ts || exit 1
-
-echo 'Compiling Headlight'
-node_modules/typescript/bin/tsc -p ./ || exit 1
 
 echo 'Compiling tests'
 node_modules/typescript/bin/tsc -p ./tests/unit || exit 1
